feat!: remove `Result#apply()`
This method is hard to grasp, hard to use.
It makes psalm cry, it makes me cry.

Since it's not even part of the `Rust` counterpart API, just remove it.

If some people wants it, they'll have to backport it into their own project and deal with the maintenance of it.